+++ packages/celestials/stars/src/main-sequence/class-m.ts
@@ -44,8 +44,11 @@ export class ClassMStarMaterial extends BaseStarMaterial {
  * Renderer for M-class stars
  */
 export class ClassMStarRenderer extends MainSequenceStarRenderer<ClassMStarMaterial> {
-  constructor(options?: BaseCelestialRendererOptions) {
-    super(options);
+  constructor(
+    object: RenderableCelestialObject,
+    options: BaseCelestialRendererOptions = {},
+  ) {
+    super(object, options);
   }
 
   /**